+++ Source/AbyssDiverUnderWorld/Monster/BT/Commons/BTTask_StopMovement.cpp
@@ -1,4 +1,4 @@
-#include "Monster/BT/AlienShark/BTTask_StopMovement.h"
+#include "Monster/BT/Commons/BTTask_StopMovement.h"
 
 #include "Monster/Monster.h"
 #include "Monster/Components/AquaticMovementComponent.h"
@@ -12,10 +12,6 @@ UBTTask_StopMovement::UBTTask_StopMovement()
 
 EBTNodeResult::Type UBTTask_StopMovement::ExecuteTask(UBehaviorTreeComponent& Comp, uint8* NodeMemory)
 {
-	// 기존 AIController의 이동 정지
-	Comp.GetAIOwner()->StopMovement();
-	
-	// AquaticMovementComponent의 이동도 정지
 	AMonster* Monster = Cast<AMonster>(Comp.GetAIOwner()->GetCharacter());
 	if (Monster && Monster->AquaticMovementComponent)
 	{
